[Sdk] move model generation to SDK - prepartions, refactoring before move #22
