+++ .gitignore
@@ -1,5 +1,6 @@
 .tmp
 .idea
 test/e2e/generated
+src/test/generated
 lib
 node_modules
