+++ README.md
@@ -77,12 +77,21 @@ npx gen-cuba-front
 ```
 
 ## Testing
+
+###Test folders
+
+/fixtures - initial data required for tests<br>
+/generated - result of generators work - apps and sdk will be stored here<br>
+/expected - files gauges used for comparison with generated code<br>
+
+
+###Run tests
 ```bash
 npm test
 ```
 
 ## Integration tests
-Integration tests are used compiled version of front-generator. To apply your code changes you need to run ```npm run compile``` before testing.
+Integration tests are used compiled version of front-generator. To apply your code changes you need to run ```npm run build``` before testing.
 <br>
 Apps and sdk generated in ```./test/e2e/generated``` directory.
 ### Run all e2e tests
