+++ package.json
@@ -10,16 +10,15 @@
     "node": ">=8.0"
   },
   "scripts": {
-    "compile": "tsc",
     "build": "rimraf lib && tsc && node scripts/copy-templates.js",
     "watch": "tsc -w",
     "test": "mocha --config test/mocha.js.opts",
 
-    "test:e2e": "mocha 'test/e2e/*.test.js' --grep 'polymer2|sdk|react' --timeout 0",
-    "test:e2e:sdk": "mocha 'test/e2e/sdk.test.js'  -g 'sdk' --timeout 0",
-    "test:e2e:react": "mocha 'test/e2e/react-client.test.js'  -g 'react' --timeout 0",
-    "test:e2e:polymer2": "mocha 'test/e2e/polymer2.test.js'  -g 'polymer2' --timeout 0",
-    "test:e2e:polymer2-ts": "mocha 'test/e2e/polymer2-ts.test.js'  -g 'polymer2-ts' --timeout 0",
+    "test:e2e": "mocha 'test/e2e/*.test.js' -g 'test:e2e:polymer2|test:e2e:sdk|test:e2e:react' --timeout 0",
+    "test:e2e:sdk": "mocha 'test/e2e/sdk.test.js'  -g 'test:e2e:sdk' --timeout 0",
+    "test:e2e:react": "mocha 'test/e2e/react-client.test.js'  -g 'test:e2e:react' --timeout 0",
+    "test:e2e:polymer2": "mocha 'test/e2e/polymer2.test.js'  -g 'test:e2e:polymer2' --timeout 0",
+    "test:e2e:polymer2-ts": "mocha 'test/e2e/polymer2-ts.test.js'  -g 'test:e2e:polymer2-ts' --timeout 0",
 
     "prepublishOnly": "npm run build"
   },
