+++ src/generators/sdk/README.md
@@ -6,6 +6,13 @@ rest [services](https://doc.cuba-platform.com/manual-latest/rest_api_v2_services
 and [queries](https://doc.cuba-platform.com/manual-latest/rest_api_v2_queries_config.html) 
 as Typescript classes.
 
+SDK could be generated in different configurations depends on option passed to ```gen-cuba-front``` command.
+
+```gen-cuba-front sdk:model``` generates entities and enums<br>
+```gen-cuba-front sdk:services``` generates rest services<br>
+```gen-cuba-front sdk:queries``` generates rest queries<br>
+```gen-cuba-front sdk:all``` generates all toolkit - entities, enums, queries and services<br>
+
 ## Entities
 
 ### Non persistent entities
