+++ src/test/fixtures/enums-model--identical-names.json
